[FIX] macom_website_sale: Perfomance and user interface
The client is not satisfied with the performance of the site, so the following changes have been made to improve the performance of the site according to the mentioned by the client:

As the page takes a long time to load, it has been chosen to avoid that all the products be loaded at the beginning of it, so they will have to resort to the search box to display them.

As the search fields are too slow, it has been decided to reduce the load generated by the methods when are filtering the information.

To improve the appearance of the page now appears in red the rows when the products are unstocked.

Navigation with keyboard arrows has also been enabled.

In the other hand, the SKU code of the product was added to the table in cart checkout.

Now the user can select a droplist item to search a product